+++ electron.vite.config.ts
@@ -7,7 +7,15 @@ export default defineConfig({
     plugins: [externalizeDepsPlugin()]
   },
   preload: {
-    plugins: [externalizeDepsPlugin()]
+    plugins: [externalizeDepsPlugin()],
+    build: {
+      rollupOptions: {
+        input: {
+          index: path.resolve(__dirname, 'src/preload/index.ts'),
+          overlay: path.resolve(__dirname, 'src/preload/overlay.ts')
+        }
+      }
+    }
   },
   renderer: {
     root: 'src/renderer',